Video Url Grabber
on a page which loads the video somehow through an iframe, there are 2 buttons: one directly on the video, second one at bottom of page.

what i want is that only the video at bottom of page exists and NOT the button on video.

obviously, when clicking the button on bottom of page, the video inside iframe should be detected